refactor(auth): 일부 패키지 이동 및 명명 수정
