+++ hongit-api/src/main/kotlin/com/devfloor/hongit/api/auth/exception/InvalidAuthException.kt
@@ -1,3 +1,3 @@
 package com.devfloor.hongit.api.auth.exception
 
-class InvalidAuthException(override val message: String) : RuntimeException()
+class InvalidAuthException(override val message: String) : RuntimeException(message)
